requirements.txt
aiobotocore==2.7.0
aiodynamo==23.10.1
fastapi==0.95.1
faster-whisper==0.9.0
openai==1.3.5
pydantic==1.10.7
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
tiktoken==0.5.2
torch==2.1.0
torchaudio==2.1.0
twilio==8.9.0
uvicorn==0.22.0
uvloop==0.19.0
websockets==10.4
python-Levenshtein
